Thorlabs: Prevent VC++ warning.
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@10942 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ DeviceAdapters/Thorlabs/IntegratedFilterWheel.cpp
@@ -551,7 +551,7 @@ int IntegratedFilterWheel::GoToPosition(long pos)
    unsigned int stepsReported = *((unsigned int*)(answer + 8));
    long posReported = (long) ((double)stepsReported / posSize + 0.5);
 
-   if (pos != stepsReported)
+   if (pos != (int)stepsReported)
    {
       ostringstream err;
       err << "GoToPosition() failed: steps requested = " << steps << ", steps reported = " << stepsReported;
